pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Yuanhao's Gemini MCP project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "mcp>=1.11.0",
    "python-dotenv>=1.0.0",
    "google-genai",
    "nest-asyncio",
    "httpx[socks]>=0.28.1",
    "langchain-google-genai",
    "langchain-core",
    "langchain",
    "langgraph"
]
